requirements.txt
streamlit==1.51.0
openai==2.7.2
requests==2.32.5
python-dotenv==1.0.1
setuptools==75.0.0
watchdog==5.0.3
python-dateutil==2.8.2
pytz==2024.1
packaging==24.1
google-cloud-secret-manager==2.20.0
Pillow==10.4.0